test bls multiple inputs
